refactor: dont override bricks via sc. part 7
