--- src/components/TableRowItem/TableRowItem.module.css
@@ -1,15 +1,0 @@
-.TableRowItem {
-    display: grid;
-    grid-template-columns: minmax(400px, 1fr) 780px;
-    align-items: center;
-    flex: 1;
-
-    gap: var(--gap-s);
-    color: var(--text-color);
-    padding: var(--gap-s);
-    border-radius: var(--radius-m);
-}
-
-.TableRowItem:hover {
-    background-color: var(--gray4);
-}
